+++ cli/src/commands/status.rs
@@ -15,10 +15,7 @@
 use itertools::Itertools as _;
 use jj_lib::copies::CopyRecords;
 use jj_lib::merge::Diff;
-use jj_lib::merged_tree::MergedTree;
 use jj_lib::repo::Repo as _;
-use jj_lib::repo_path::RepoPath;
-use jj_lib::repo_path::RepoPathBuf;
 use jj_lib::revset::RevsetExpression;
 use jj_lib::revset::RevsetFilterPredicate;
 use pollster::FutureExt as _;
@@ -28,6 +25,7 @@ use crate::cli_util::CommandHelper;
 use crate::cli_util::print_conflicted_paths;
 use crate::cli_util::print_snapshot_stats;
 use crate::cli_util::print_unmatched_explicit_paths;
+use crate::cli_util::visit_collapsed_untracked_files;
 use crate::command_error::CommandError;
 use crate::diff_util::DiffFormat;
 use crate::diff_util::get_copy_records;
@@ -55,6 +53,9 @@ pub(crate) struct StatusArgs {
     /// Restrict the status display to these paths
     #[arg(value_name = "FILESETS", value_hint = clap::ValueHint::AnyPath)]
     paths: Vec<String>,
+    /// Show ignored files as well.
+    #[arg(short, long, action = clap::ArgAction::SetTrue)]
+    ignored: bool,
 }
 
 #[instrument(skip_all)]
@@ -64,11 +65,6 @@ pub(crate) fn cmd_status(
     args: &StatusArgs,
 ) -> Result<(), CommandError> {
     let (workspace_command, snapshot_stats) = command.workspace_helper_with_stats(ui)?;
-    print_snapshot_stats(
-        ui,
-        &snapshot_stats,
-        workspace_command.env().path_converter(),
-    )?;
     let repo = workspace_command.repo();
     let maybe_wc_commit = workspace_command
         .get_wc_commit_id()
@@ -83,11 +79,17 @@ pub(crate) fn cmd_status(
     if let Some(wc_commit) = &maybe_wc_commit {
         let parent_tree = wc_commit.parent_tree(repo.as_ref())?;
         let tree = wc_commit.tree();
+        print_snapshot_stats(
+            ui,
+            &snapshot_stats,
+            workspace_command.env().path_converter(),
+        )?;
 
         print_unmatched_explicit_paths(ui, &workspace_command, &fileset_expression, [&tree])?;
 
         let wc_has_changes = tree.tree_ids() != parent_tree.tree_ids();
         let wc_has_untracked = !snapshot_stats.untracked_paths.is_empty();
+        let wc_has_ignored = !snapshot_stats.ignored_paths.is_empty();
         if !wc_has_changes && !wc_has_untracked {
             writeln!(formatter, "The working copy has no changes.")?;
         } else {
@@ -116,7 +118,7 @@ pub(crate) fn cmd_status(
                 writeln!(formatter, "Untracked paths:")?;
                 visit_collapsed_untracked_files(
                     snapshot_stats.untracked_paths.keys(),
-                    tree.clone(),
+                    &tree,
                     |path, is_dir| {
                         let ui_path = workspace_command.path_converter().format_file_path(path);
                         writeln!(
@@ -132,6 +134,22 @@ pub(crate) fn cmd_status(
                     },
                 )
                 .block_on()?;
+            }
+        }
+
+        if args.ignored && wc_has_ignored {
+            writeln!(formatter, "Ignored paths:")?;
+            for (path, is_dir) in snapshot_stats.ignored_paths {
+                let ui_path = workspace_command.path_converter().format_file_path(&path);
+                writeln!(
+                    formatter.labeled("diff").labeled("ignored"),
+                    "! {ui_path}{}",
+                    if is_dir {
+                        std::path::MAIN_SEPARATOR_STR
+                    } else {
+                        ""
+                    }
+                )?;
             }
         }
 
@@ -231,137 +249,4 @@ pub(crate) fn cmd_status(
     }
 
     Ok(())
-}
-
-async fn visit_collapsed_untracked_files(
-    untracked_paths: impl IntoIterator<Item = impl AsRef<RepoPath>>,
-    tree: MergedTree,
-    mut on_path: impl FnMut(&RepoPath, bool) -> Result<(), CommandError>,
-) -> Result<(), CommandError> {
-    let trees = tree.trees().await?;
-    let mut stack = vec![trees];
-
-    // TODO: This loop can be improved with BTreeMap cursors once that's stable,
-    // would remove the need for the whole `skip_prefixed_by` thing and turn it
-    // into a B-tree lookup.
-    let mut skip_prefixed_by_dir: Option<RepoPathBuf> = None;
-    'untracked: for path in untracked_paths {
-        let path = path.as_ref();
-        if skip_prefixed_by_dir
-            .as_ref()
-            .is_some_and(|p| path.starts_with(p))
-        {
-            continue;
-        } else {
-            skip_prefixed_by_dir = None;
-        }
-
-        let mut it = path.components().dropping_back(1);
-        let first_mismatch = it.by_ref().enumerate().find(|(i, component)| {
-            stack.get(i + 1).is_none_or(|tree| {
-                tree.dir()
-                    .components()
-                    .next_back()
-                    .expect("should always have at least one element (the root)")
-                    != *component
-            })
-        });
-
-        if let Some((i, component)) = first_mismatch {
-            stack.truncate(i + 1);
-            for component in std::iter::once(component).chain(it) {
-                let parent = stack
-                    .last()
-                    .expect("should always have at least one element (the root)");
-
-                if let Some(subtree) = parent.sub_tree(component).await? {
-                    stack.push(subtree);
-                } else {
-                    let dir = parent.dir().join(component);
-
-                    on_path(&dir, true)?;
-                    skip_prefixed_by_dir = Some(dir);
-
-                    continue 'untracked;
-                }
-            }
-        }
-
-        on_path(path, false)?;
-    }
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod test {
-    use testutils::TestRepo;
-    use testutils::TestTreeBuilder;
-    use testutils::repo_path;
-
-    use super::*;
-
-    fn collect_collapsed_untracked_files_string(
-        untracked_paths: &[&RepoPath],
-        tree: MergedTree,
-    ) -> String {
-        let mut result = String::new();
-        visit_collapsed_untracked_files(untracked_paths, tree, |path, is_dir| {
-            result.push_str("? ");
-            if is_dir {
-                result.push_str(&path.to_internal_dir_string());
-            } else {
-                result.push_str(path.as_internal_file_string());
-            }
-            result.push('\n');
-            Ok(())
-        })
-        .block_on()
-        .unwrap();
-        result
-    }
-
-    #[test]
-    fn test_collapsed_untracked_files() {
-        let repo = TestRepo::init();
-
-        let tracked = {
-            let mut builder = TestTreeBuilder::new(repo.repo.store().clone());
-
-            builder.file(repo_path("top_level_file"), "");
-            // ? "untracked_top_level_file"
-            // ? "dir"
-            // ? "dir2/c"
-            builder.file(repo_path("dir2/d"), "");
-            // ? "dir3/partially_tracked/e"
-            builder.file(repo_path("dir3/partially_tracked/f"), "");
-            // ? "dir3/fully_untracked/"
-            builder.file(repo_path("dir3/j"), "");
-            // ? "dir3/k"
-
-            builder.write_merged_tree()
-        };
-        let untracked = &[
-            repo_path("untracked_top_level_file"),
-            repo_path("dir/a"),
-            repo_path("dir/b"),
-            repo_path("dir2/c"),
-            repo_path("dir3/partially_tracked/e"),
-            repo_path("dir3/fully_untracked/g"),
-            repo_path("dir3/fully_untracked/h"),
-            repo_path("dir3/k"),
-        ];
-
-        insta::assert_snapshot!(
-            collect_collapsed_untracked_files_string(untracked, tracked),
-            @r"
-        ? untracked_top_level_file
-        ? dir/
-        ? dir2/c
-        ? dir3/partially_tracked/e
-        ? dir3/fully_untracked/
-        ? dir3/k
-        "
-        );
-    }
 }